backend_app/requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework
gunicorn
sqlparse==0.5.3
